feat: dio singleton 적용
