+++ lib/core/network/api_client.dart
@@ -1,0 +1,40 @@
+import 'package:dio/dio.dart';
+import 'package:flutter/foundation.dart';
+import 'package:pretty_dio_logger/pretty_dio_logger.dart';
+
+import '../constants/constants.dart';
+
+class ApiClient {
+  static final ApiClient _instance = ApiClient._internal();
+
+  factory ApiClient() => _instance;
+
+  late final Dio dio;
+
+  ApiClient._internal() {
+    dio = Dio(
+      BaseOptions(
+        baseUrl: BASE_URL,
+        connectTimeout: const Duration(seconds: 5),
+        receiveTimeout: const Duration(seconds: 3),
+        headers: {
+          'Authorization': ACCESS_TOKEN,
+          'X-Company-Code': X_COMPANY_CODE,
+        },
+      ),
+    );
+
+    dio.interceptors.add(
+      PrettyDioLogger(
+        requestHeader: true,
+        requestBody: true,
+        responseBody: true,
+        responseHeader: false,
+        error: true,
+        compact: true,
+        maxWidth: 90,
+        enabled: kDebugMode,
+      ),
+    );
+  }
+}
