+++ lib/service/product_api/login_service.dart
@@ -1,10 +1,12 @@
 import 'package:dio/dio.dart';
 
-import '../api_client.dart';
+import '../../core/network/api_client.dart';
 
 class LogInService {
+  final dio = ApiClient().dio;
+
   Future<Response> signIn(String email, String password) async {
-    return await ApiClient.dio.post('/api/auth/login', data: {
+    return await dio.post('/api/auth/login', data: {
       'email': email,
       'password': password,
     });
