+++ lib/service/product_api/product_api.dart
@@ -1,8 +1,10 @@
 import 'package:dio/dio.dart';
-import '../api_client.dart';
+import '../../core/network/api_client.dart';
 
 class ProductService {
+  final dio = ApiClient().dio;
+
   Future<Response> getProducts() async {
-    return await ApiClient.dio.get('/api/products');
+    return await dio.get('/api/products');
   }
 }
